requirements.txt
appdirs==1.4.0
Django==1.10.5
django-crispy-forms==1.6.1
django-taggit==0.22.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
six==1.10.0
